requirements.txt
Django==5.1.1
djangorestframework==3.15.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
django-jazzmin==3.0.0